YEL-24 [refactor] ImmutableMap -> ServiceTokenVO 변경
