+++ build.gradle
@@ -70,9 +70,6 @@ dependencies {
     implementation "com.querydsl:querydsl-jpa:${queryDslVersion}"
     implementation "com.querydsl:querydsl-apt:${queryDslVersion}"
 
-    // For ImmutableMap
-    implementation 'com.google.guava:guava:31.1-jre'
-
     runtimeOnly 'com.h2database:h2'
     runtimeOnly 'com.mysql:mysql-connector-j:8.0.31'
 }
